Изменен параметр x в мтеодах (Integer -> Double)

+++ src/main/java/coordinateProcessing/CartesianAreas.java
@@ -1,9 +1,9 @@
 package coordinateProcessing;
 
 public interface CartesianAreas {
-    boolean firstRotation(Integer x, Double y, Float R);
-    boolean secondRotation(Integer x, Double y, Float R);
-    boolean thirdRotation(Integer x, Double y, Float R);
-    boolean fourthRotation(Integer x,Double y,Float R);
-    boolean areaCheck(Integer x,Double y,Float R);
+    boolean firstRotation(Double x, Double y, Float R);
+    boolean secondRotation(Double x, Double y, Float R);
+    boolean thirdRotation(Double x, Double y, Float R);
+    boolean fourthRotation(Double x,Double y,Float R);
+    boolean areaCheck(Double x,Double y,Float R);
 }
